# box/kubernetes/forklift/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: forklift API
  version: "0.1.0"
  description: >
    Management API for forklift, a Kubernetes-native artifact repository.
    Package-format endpoints (Maven, npm, Cargo, Go, PyPI, raw) are mounted under
    /maven, /npm, /cargo, /go, /pypi and /raw and follow each ecosystem's native
    protocol (raw serves plain GET/PUT by path). The OCI distribution API
    (container images, Helm charts) is mounted at /v2 with the repository name
    as the first path segment. This specification is served at
    /openapi.yaml with a Scalar reference UI at /api-docs.
servers:
  - url: /
...
paths:
...
  /api/v1/announcement:
    get:
      tags: [session]
      summary: Read the site-wide announcement banner
      description: >
        Readable by any signed-in user. Never set and cleared are the same
        response - an empty body - so the banner has one absent state to render
        rather than two.
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: The current announcement, or an empty body if there is none
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
    put:
      tags: [session]
      summary: Replace the site-wide announcement (admin)
      description: >
        The body is stored as Markdown source verbatim; rendering and
        sanitisation are the UI's job. An empty body clears the announcement.
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "200":
          description: The stored announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400":
          description: Body exceeds 1000 characters or 16KiB
...
  /api/v1/users:
    get:
      tags: [users]
      summary: List users (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      tags: [users]
      summary: Create a local user or robot account (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: The created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedUser" }
        "400": { description: Validation error (e.g. a robot with a password, or a user without one) }
        "409": { description: Username already exists }
  /api/v1/users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      tags: [users]
      summary: Update a user (admin) - reset password, toggle disabled or lockout, or unlock
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "200":
          description: The updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Validation error (incl. disabling your own account) }
        "404": { description: Not found }
    delete:
      tags: [users]
      summary: Delete a user (admin); deleting your own account is rejected
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "400": { description: Cannot delete your own account }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Local username/password or a personal access token as the password.
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token (flpat_...) or an OIDC bearer token.
...
  schemas:
    RoleRef:
      type: object
      description: A role as it appears inside a user record, without its permissions.
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...
    User:
      type: object
      required: [created_at, disabled, email, failed_login_count, id, last_login_at, locked, lockout_enabled, protected, robot, roles, source, token_count, username]
      properties:
        id: { type: integer }
        username: { type: string }
        source: { type: string, enum: [local, oidc] }
        email: { type: string }
        disabled: { type: boolean }
        robot:
          type: boolean
          description: A token-only service account, which cannot log in interactively.
        created_at: { type: string, format: date-time }
        last_login_at:
          type: [string, "null"]
          format: date-time
          description: Null when the user has never logged in.
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleRef" }
        lockout_enabled: { type: boolean }
        locked: { type: boolean }
        failed_login_count:
          type: integer
          description: Consecutive failed password logins; the account locks at the hard limit of 5. Reset on success or unlock.
        protected:
          type: boolean
          description: The bootstrap admin, which cannot be locked out.
        token_count:
          type: integer
          description: Personal access tokens the user owns, so list views need no extra request.
...
    Announcement:
      type: object
      description: >
        The site-wide banner. An empty body means there is nothing to show;
        updated_by is absent until someone has set one.
      required: [body, updated_at]
      properties:
        body: { type: string, description: Markdown source, rendered by the UI. }
        updated_by: { type: string }
        updated_at: { type: string, format: date-time }
    AnnouncementInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          description: >
            Markdown source. Capped at 1000 characters and 16KiB; an empty
            string clears the announcement.
...
    CreatedUser:
      type: object
      description: >
        The slim acknowledgement of a created user - enough to address the new
        row. The full record is a separate GET.
      required: [id, username]
      properties:
        id: { type: integer }
        username: { type: string }
...
    UserCreate:
      type: object
      required: [username]
      properties:
        username: { type: string }
        password:
          type: string
          description: Required for a normal user; must be omitted for a robot account.
        email: { type: string }
        role_ids: { type: array, items: { type: integer } }
        robot:
          type: boolean
          description: >
            Create a token-only Robot Account: no password, interactive
            login (password and OIDC session) is refused, but the account
            can own personal access tokens for automation.
    UserInput:
      type: object
      properties:
        password: { type: string, description: New password (local users only). }
        disabled: { type: boolean }
        lockout_enabled: { type: boolean, description: Enable or disable account lockout after repeated failed logins. }
        unlock: { type: boolean, description: Clear an active lockout immediately. }
...